requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.2
